[REFACTOR] ViewControllerable 폴더 정리
