+++ LionHeart-iOS/LionHeart-iOS/Scenes/Article/ArticleCategory/ViewControllers/ArticleCategoryViewController.swift
@@ -10,10 +10,6 @@ import UIKit
 
 import SnapKit
 
-protocol ArticleCategoryViewControllerable where Self: UIViewController {
-    var coordinator: ArticleCategoryNavigation? {get set}
-}
-
 final class ArticleCategoryViewController: UIViewController, ArticleCategoryViewControllerable {
     
     weak var coordinator: ArticleCategoryNavigation?
