+++ LionHeart-iOS/LionHeart-iOS/Scenes/BookMark/ViewControllers/BookmarkViewController.swift
@@ -10,10 +10,6 @@ import UIKit
 
 import SnapKit
 
-protocol BookmarkViewControllerable where Self: UIViewController {
-    var coordinator: BookmarkNavigation? { get set }
-}
-
 final class BookmarkViewController: UIViewController, BookmarkViewControllerable {
     
     weak var coordinator: BookmarkNavigation?
