+++ LionHeart-iOS/LionHeart-iOS/Scenes/Coordinator/Interface/ArticleCategoryViewControllerable.swift
@@ -1,0 +1,12 @@
+//
+//  ArticleCategoryViewControllerable.swift
+//  LionHeart-iOS
+//
+//  Created by uiskim on 2023/10/15.
+//
+
+import UIKit
+
+protocol ArticleCategoryViewControllerable where Self: UIViewController {
+    var coordinator: ArticleCategoryNavigation? {get set}
+}
