+++ LionHeart-iOS/LionHeart-iOS/Scenes/Coordinator/Interface/BookmarkViewControllerable.swift
@@ -1,0 +1,12 @@
+//
+//  BookmarkViewControllerable.swift
+//  LionHeart-iOS
+//
+//  Created by uiskim on 2023/10/15.
+//
+
+import UIKit
+
+protocol BookmarkViewControllerable where Self: UIViewController {
+    var coordinator: BookmarkNavigation? { get set }
+}
