+++ LionHeart-iOS/LionHeart-iOS/Scenes/Coordinator/Interface/ChallengeViewControllerable.swift
@@ -1,0 +1,12 @@
+//
+//  ChallengeViewControllerable.swift
+//  LionHeart-iOS
+//
+//  Created by uiskim on 2023/10/15.
+//
+
+import UIKit
+
+protocol ChallengeViewControllerable where Self: UIViewController {
+    var coordinator: ChallengeNavigation? { get set }
+}
